requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
av==14.2.0
click==8.1.8
cmake==3.31.6
colorama==0.4.6
distlib==0.3.9
dlib==19.24.6
exceptiongroup==1.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.115.8
filelock==3.17.0
h11==0.14.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
libarchive-c==5.2
numpy==2.0.2
opencv-python==4.11.0.86
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pillow==11.1.0
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-prctl==1.8.1
referencing==0.36.2
rpds-py==0.23.1
rpi-libcamera==0.1a9
simplejpeg==1.8.2
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
v4l2-python3==0.3.5
virtualenv==20.29.2
